ml-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4

groq==0.13.0

langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface==0.0.3

sentence-transformers==2.6.1
faiss-cpu==1.7.4

numpy==1.26.4
httpx==0.27.0
requests==2.31.0